agregando modificaciones a cartera

+++ add-edit-delete.php
@@ -70,13 +70,13 @@ echo json_encode(true);
     if($_POST['mode'] === 'edit_cartera'){
         //$result = mysqli_query($conn,"SELECT * FROM cartera WHERE id='" . $_POST['id'] . "'");
         //$result = mysqli_query($conn,"SELECT c.id, c.fecha, c.persona_fk, concat(p.nombre,' ',p.apellido) as nombre, c.local_fk, l.numero as num_local, c.valor, c.servicio, c.pendiente FROM cartera c INNER JOIN locales l on c.local_fk = l.id INNER JOIN personas p on c.persona_fk = p.id WHERE c.id =  '". $_POST['id']."'");
-        $result = mysqli_query($conn,"SELECT i.id, i.fecha_desde,i.fecha_hasta,i.fecha_ingreso, i.persona_fk, concat(p.nombre,' ',p.apellido) as nombre, i.local_fk, l.numero as num_local, i.recibo,i.categoria,i.sub_categoria,i.valor,i.pendiente,i.observaciones FROM ingresos i INNER JOIN locales l on i.local_fk = l.id INNER JOIN personas p on i.persona_fk = p.id WHERE i.id =  '". $_POST['id']."'");
+        $result = mysqli_query($conn,"SELECT i.id, i.fecha_desde,i.fecha_hasta,i.fecha_ingreso, i.fecha_pago, i.persona_fk, concat(p.nombre,' ',p.apellido) as nombre, i.local_fk, l.numero as num_local, i.recibo,i.categoria,i.sub_categoria,i.valor,i.pendiente,i.observaciones FROM ingresos i INNER JOIN locales l on i.local_fk = l.id INNER JOIN personas p on i.persona_fk = p.id WHERE i.id =  '". $_POST['id']."'");
         $row= mysqli_fetch_array($result);
         echo json_encode($row);
     }
 
     if ($_POST['mode'] === 'update_cartera') {
-        mysqli_query($conn,"UPDATE ingresos set  fecha_desde='" . $_POST['fecha_desde'] . "', fecha_hasta='" . $_POST['fecha_hasta'] . "', fecha_ingreso='" . $_POST['fecha_ing'] . "', persona_fk='" . $_POST['perteneces_id'] . "', local_fk='" . $_POST['pertenece_id'] . "', recibo='" . $_POST['recibo'] ."', categoria='" . $_POST['categoria'] . "', sub_categoria='" . $_POST['sub_categ'] . "', valor='" . $_POST['valor'] . "', pendiente='" . $_POST['pendiente'] . "', observaciones='" . $_POST['obs'] . "' WHERE id='" . $_POST['id'] . "'");
+        mysqli_query($conn,"UPDATE ingresos set  fecha_desde='" . $_POST['fecha_desde'] . "', fecha_hasta='" . $_POST['fecha_hasta'] . "', fecha_ingreso='" . $_POST['fecha_ing'] . "', fecha_pago='" . $_POST['fecha_pago'] . "', persona_fk='" . $_POST['perteneces_id'] . "', local_fk='" . $_POST['pertenece_id'] . "', recibo='" . $_POST['recibo'] ."', categoria='" . $_POST['categoria'] . "', sub_categoria='" . $_POST['sub_categ'] . "', valor='" . $_POST['valor'] . "', pendiente='" . $_POST['pendiente'] . "', observaciones='" . $_POST['obs'] . "' WHERE id='" . $_POST['id'] . "'");
         //mysqli_query($conn,"UPDATE cartera set  fecha='" . $_POST['fecha'] . "', persona_fk='" . $_POST['perteneces_id'] . "', local_fk='" . $_POST['pertenece_id'] . "', valor='" . $_POST['valor'] . "', servicio='" . $_POST['servicio'] . "', pendiente='" . $_POST['pendiente'] ."' WHERE id='" . $_POST['id'] . "'");
         echo json_encode(true);
     } 
@@ -102,9 +102,18 @@ echo json_encode(true);
 
 
     if ($_POST['mode'] === 'delete_cartera') {
-        mysqli_query($conn, "DELETE FROM ingresos WHERE id='" . $_POST["id"] . "'");
-        $registro_delete = mysqli_query($conn, "DELETE FROM cartera WHERE id='" . $_POST["id"] . "'");
-        echo json_encode(true);
+       //debemos mirar si ese ingreso contiene un abono para que si lo contiene el usuario primero debe borrar dicho abono
+
+       $sql = "select ingresos_fk from abonos where ingresos_fk = '".$_POST["id"]."'";
+       $res = mysqli_query($conn,$sql);
+
+       if(mysqli_num_rows($res) == 0){
+           mysqli_query($conn, "DELETE FROM ingresos WHERE id='" . $_POST["id"] . "'");
+           $registro_delete = mysqli_query($conn, "DELETE FROM ingresos WHERE id='" . $_POST["id"] . "'");
+           echo json_encode(true);
+       }else{
+           echo json_encode("Este item contiene un abono, debe borrar primero el abono para borrar este registro");
+       }
     }
 
     //pagar cartera
